backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
sqlalchemy==2.0.28
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.6.4
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1
langserve[server]==0.0.51
langchain==0.1.12
langchain-openai==0.0.8
aiosmtplib==3.0.1
mypy==1.9.0